requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.0.5
django-cors-headers==4.0.0
django-countries==7.5.1
djangorestframework==3.13.1
idna==3.4
Pillow==9.5.0
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
typing-extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2